requirements.txt
#WEB SCRAPING
bs4
selenium
scrapy
html5lib

#DATABASE
pymongo
mysql-connector-python

#UTILITIES
colorama
python-dotenv